home-cloud-backend/backend/requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
pydantic==2.1.1
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.1
